requirements.txt
boto3==1.24.88
botocore==1.27.88
certifi==2022.9.24
charset-normalizer==2.1.1
docopt==0.6.2
idna==3.4
jmespath==1.0.1
mysql-connector-python==8.0.30
pipreqs==0.4.11
protobuf==3.20.1
python-dateutil==2.8.2
requests==2.28.1
s3transfer==0.6.0
urllib3==1.26.12
yarg==0.1.9
flask
